add class diagram
